feat(api): seederService injectable
