+++ packages/api/src/User/User.entity.ts
@@ -5,6 +5,7 @@ import {
   Entity,
   PrimaryGeneratedColumn,
   Unique,
+  BaseEntity
 } from 'typeorm';
 
 @ObjectType()
@@ -27,7 +28,7 @@ export class User {
 
 @Entity('user')
 @Unique('Email', ['email'])
-export class UserWithPassword implements User {
+export class UserWithPassword extends BaseEntity {
   @PrimaryGeneratedColumn('uuid')
   id: string;
